Reorganised files and folders into layered architecture pattern
