+++ codar-web/presentation/shared_presentation/sidepanel.php
